// src/generated/opentofu-api.ts
export interface definitions {
// ...
  Module: {
    addr: definitions["ModuleAddr"];
    blocked_reason?: string;
    description: string;
    /**
     * Format: int64
     * @description ForkCount indicates how many forks this provider has.
     */
    fork_count: number;
    fork_of?: definitions["ModuleAddr"];
    /** @description ForkOfLink may contain a link to a repository this provider is forked from. */
    fork_of_link?: string;
    is_blocked: boolean;
    /**
     * Format: int64
     * @description Popularity indicates how popular the underlying repository is in the VCS system.
     */
    popularity: number;
    /**
     * Format: int64
     * @description UpstreamForkCount contains the number of forks of the upstream repository.
     */
    upstream_fork_count?: number;
    /**
     * Format: int64
     * @description UpstreamPopularity contains the popularity of the original repository this repository is forked of.
     */
    upstream_popularity?: number;
    versions: definitions["ModuleVersionDescriptor"][];
  };
// ...
  ProviderDocItem: {
    description?: string;
    edit_link?: string;
    name?: definitions["DocItemName"];
    subcategory?: string;
    title?: string;
  };
// ...
  SearchResultItem: {
    /** @description The address of the module or provider. */
    addr: string;
    /** @description A brief description of the result item. */
    description: string;
    /** @description The unique identifier for the result item. */
    id: string;
    /** @description The last updated timestamp for the result item. */
    last_updated: string;
    /** @description A map of variables used to generate the link for the result item. */
    link_variables: { [key: string]: string };
    /**
     * Format: int32
     * @description The rank of the result in the search results.
     */
    rank: number;
    /** @description The number of times the search term matched in this result. */
    term_match_count: string;
    /** @description The title of the result item. */
    title: string;
    /** @description The type of the result item (e.g., module, provider, datasource etc). */
    type: string;
    /** @description The version of the module or provider */
    version: string;
  };
// ...
}